+++ Threeyes/Plugins/ThreeyesPluginExpert/Module/Steamworks/Hub/Simulator/SceneManagerSimulator.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Threeyes.Core;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using System;
 
 #if UNITY_EDITOR
 using UnityEditor;
@@ -12,11 +13,17 @@ namespace Threeyes.Steamworks
 {
     public class SceneManagerSimulator : HubSceneManagerBase<SceneManagerSimulator>
     {
-        public bool HasSceneLoaded { get { return hasSceneLoaded; } }
-        bool hasSceneLoaded = false;
+        //public bool HasSceneLoaded { get { return hasSceneLoaded; } }
+        //bool hasSceneLoaded = false;
 
         [SerializeField] protected SOWorkshopItemInfo curSOWorkshopItemInfo;
         [SerializeField] protected WorkshopItemInfo curWorkshopItemInfo;
+
+        protected override void SetInstanceFunc()
+        {
+            base.SetInstanceFunc();
+            isChangingScene = true;//提前标记，避免其他组件提前初始化
+        }
         protected virtual void Start()
         {
             InitAsync();
@@ -40,6 +47,7 @@ namespace Threeyes.Steamworks
             if (!curModScene.IsValid())
             {
                 Debug.LogError("Please add the Mod Scene before play!");
+                isChangingScene = false;
                 return;
             }
             //#1 扫描获取对应的WorkshopItemInfo（实现方法：从该Scene的路径开始，往上搜索到Items文件夹，就能知道Item的名称，然后通过方法找到对应的WorkshopInfo）
@@ -52,6 +60,7 @@ namespace Threeyes.Steamworks
             if (curSOWorkshopItemInfo == null)
             {
                 Debug.LogError($"Can't find {nameof(SOWorkshopItemInfo)} for item {itemName}!");
+                isChangingScene = false;
                 return;
             }
             curWorkshopItemInfo = curSOWorkshopItemInfo.BaseItemInfo;
@@ -61,11 +70,35 @@ namespace Threeyes.Steamworks
             if (!modEntry)
             {
                 Debug.LogError($"Can't find {nameof(ModEntry)} in Mod Scene!");
+                isChangingScene = false;
                 return;
             }
+
             InitMod(modEntry);
-            hasSceneLoaded = true;
+            isChangingScene = false;
 #endif 
+        }
+
+        protected override void InitMod(ModEntry modEntry)
+        {
+            try
+            {
+                InitModFunc(modEntry);
+            }
+            catch (Exception e)
+            {
+                Debug.LogError($"Init Mod [{curWorkshopItemInfo.title}] with error: \r\n" + e);//写入到对应Log中
+            }
+        }
+
+        protected virtual void InitModFunc(ModEntry modEntry)
+        {
+            //#1 PreInit Scene Scripts
+            ManagerHolder.GetListManagerModPreInitOrder().ForEach(m => m.OnModPreInit(curModScene, modEntry));
+            EventCommunication.SendMessage<IModPreHandler>((inst) => inst.OnModPreInit());
+
+            //#2 Init Mod
+            base.InitMod(modEntry);
         }
     }
 }