+++ Threeyes/SDK/Scripts/Mod/Mod/Controller/Interface/IAD_EnvironmentController.cs
@@ -4,4 +4,6 @@ using UnityEngine;
 using Threeyes.Steamworks;
 public interface IAD_EnvironmentController : IEnvironmentController
 {
+    void RegisterCustomSunEntity(AD_SunEntityController customSunEntityController);
+    void UnRegisterCustomSunEntity(AD_SunEntityController customSunEntityController);
 }